LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-from-to-rotate-auto-expected.txt


FAIL <animateMotion> from/to animation with rotate=auto assert_approx_equals: bbox x expected 0 +/- 0.5 but got -200
